% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testbreedr.R
\name{testbreedr}
\alias{testbreedr}
\title{Test that breedR package loads and works}
\usage{
testbreedr()
}
\value{
object of "breedR" class
}
\description{
Test that breedR package loads and works
}
\examples{
  summary(testbreedr())
}
